refactor: move state reducers logic into their own files
