--- packages/store-utils/jest.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-  displayName: 'store-utils',
-  preset: '../../jest.preset.js',
-  transform: {
-    '^.+\\.[tj]sx?$': 'babel-jest',
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../coverage/packages/store-utils',
-};
