--- packages/store-utils/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@joonasmkauppinen/store-utils",
-  "version": "0.0.1"
-}
