--- packages/store-utils/project.json
@@ -1,43 +1,0 @@
-{
-  "root": "packages/store-utils",
-  "sourceRoot": "packages/store-utils/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "build": {
-      "executor": "@nrwl/web:rollup",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/packages/store-utils",
-        "tsConfig": "packages/store-utils/tsconfig.lib.json",
-        "project": "packages/store-utils/package.json",
-        "entryFile": "packages/store-utils/src/index.ts",
-        "external": ["react/jsx-runtime"],
-        "rollupConfig": "@nrwl/react/plugins/bundle-rollup",
-        "compiler": "babel",
-        "assets": [
-          {
-            "glob": "packages/store-utils/README.md",
-            "input": ".",
-            "output": "."
-          }
-        ]
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["packages/store-utils/**/*.{ts,tsx,js,jsx}"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/packages/store-utils"],
-      "options": {
-        "jestConfig": "packages/store-utils/jest.config.js",
-        "passWithNoTests": true
-      }
-    }
-  }
-}
