--- packages/store-utils/src/index.ts
@@ -1,2 +1,0 @@
-export * from './sample-data';
-export * from './sample-data/types';
