--- packages/store-utils/src/sample-data/index.ts
@@ -1,1 +1,0 @@
-export * from './sampleCardsState/sampleCardsState';
