--- packages/store-utils/src/sample-data/types/index.ts
@@ -1,2 +1,0 @@
-export * from './storeActions/storeActions';
-export * from './appState/appState';
