+++ packages/store-zustand/src/index.ts
@@ -1,1 +1,3 @@
 export * from './store/zustandStore';
+export * from './actions';
+export * from './types';
